Revamp 3/3
Most functionalities are done. Missing Transfer between Accounts.
Still a few bugs to fix and some small changes.

+++ Headers/struct.h
@@ -15,7 +15,7 @@ struct Account {
     char* IBan;
     char* id_user;
     char* coin;
-    unsigned long amount;
+    unsigned long long amount;
 
 };
